Use exact matches in asset routes
The previous way mixed the js/css routes, so that a request for
a assets/js/css_name.js and assets/css/js_name.css would be
matched.

I'm not sure it's possible to use a single regexp to match
against both while still having the correct captures yielded,
so just use two regexps in an array.