Eliminate :css_engine and :js_engine
This eliminates the engine options.  Users can just specify the file extension
when using the :js and :css options.  This makes it easy to mix and match
multiple asset engines (such as mixing scss and less in the same project).

While here, stop expanding the :path option, there shouldn't be a need for
it.  If users want an expanded path, they can specify one.

Add a default 2nd argument in the load_dependencies call, so the plugin
actually works with no arguments.

Discourage use of modifying assets_opts directly, though it is still
supported.  The main problem with accessing it after the call is that
if you modify :headers, it won't get copied to :js_headers and
:css_headers.  A similar change should probably be made to the
render plugin.