Don't include compiled extension in noncompiled mode
This allows for requests for the source extension, so for
:css=>'foo.scss', it will use /assets/css/foo.scss instead of
/assets/css/foo.scss.css.

This also refactors the routing so it just does two separate
routes, one for css and one for js, and it doesn't have to
capture the type, since it is passed in.  Also, it now only
matches GET requests, instead of all request methods.