Fix plugin options handling
Make non-compiled case not use :compiled_* options.

Handle case where plugin is called multiple times with different
options.

Try to make sure every plugin value is frozen.

No longer override inherited.  The assets_opts hash itself is
frozen, as should be all of the values in the hash, so you
shouldn't be able to modify the subclass and have the values
affect the parent hash.

Add specs for testing plugin options, specifically how the hash
values are merged and how the paths are computed.