Add :group_subdirs option, for not assuming that asset groups require subdirectories
With this option I'm envisioning an asset directory with many
shared assets, where each asset group needs a different set of the
shared assets.  When asset groups force subdirectories, that
becomes difficult to handle without symlinking.  With this option,
you can have an arbitrary number of groups, each which has a
different set of the shared assets, and compiling the assets
compiles each set to a single file.